Refactor Email and add more logging
(cherry picked from sonarr 5301620ecfd6cffb6014e986d1ce5b6bbaace93a)
closes #6168
